pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videopub"
version = "1.0.0"
description = "多平台视频发布自动化工具"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "loguru>=0.7",
    "watchdog>=3.0",
    "python-docx>=1.0",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-auth>=2.0",
    "biliup>=0.4",
    "playwright>=1.40",
    "patchright>=1.52.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
videopub = "videopub.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
